Split widget: signal table header
