+++ extension/devtools/app/lib/main.dart
@@ -40,6 +40,7 @@ part 'widgets/diff_token.dart';
 part 'widgets/metric_tile.dart';
 part 'widgets/signals_header.dart';
 part 'widgets/signal_list.dart';
+part 'widgets/signal_table_header.dart';
 part 'widgets/panel_placeholder.dart';
 part 'widgets/panel_scroll_view.dart';
 part 'widgets/panel_state_cards.dart';
@@ -2787,81 +2788,6 @@ class _EffectRow extends StatelessWidget {
           ),
         ),
       ],
-    );
-  }
-}
-
-class _SignalTableHeader extends StatelessWidget {
-  const _SignalTableHeader({
-    required this.sortKey,
-    required this.sortAscending,
-    required this.onSortName,
-    required this.onSortUpdated,
-  });
-
-  final _SortKey sortKey;
-  final bool sortAscending;
-  final VoidCallback onSortName;
-  final VoidCallback onSortUpdated;
-
-  @override
-  Widget build(BuildContext context) {
-    final labelStyle = Theme.of(context).textTheme.labelSmall;
-
-    return Container(
-      padding: const EdgeInsets.symmetric(horizontal: 16, vertical: 12),
-      decoration: BoxDecoration(
-        border: Border(
-          bottom: BorderSide(
-            color: Theme.of(context).dividerColor.withValues(alpha: 0.4),
-          ),
-        ),
-      ),
-      child: Row(
-        children: [
-          Expanded(
-            flex: 3,
-            child: _SortHeaderCell(
-              label: 'Name',
-              isActive: sortKey == _SortKey.name,
-              ascending: sortAscending,
-              onTap: onSortName,
-              style: labelStyle,
-            ),
-          ),
-          Expanded(
-            flex: 2,
-            child: Align(
-              alignment: Alignment.center,
-              child: Text('Value', style: labelStyle),
-            ),
-          ),
-          Expanded(
-            flex: 2,
-            child: Align(
-              alignment: Alignment.center,
-              child: Text('Type', style: labelStyle),
-            ),
-          ),
-          Expanded(
-            flex: 2,
-            child: Align(
-              alignment: Alignment.center,
-              child: Text('Status', style: labelStyle),
-            ),
-          ),
-          Expanded(
-            flex: 2,
-            child: _SortHeaderCell(
-              label: 'Updated',
-              isActive: sortKey == _SortKey.updated,
-              ascending: sortAscending,
-              onTap: onSortUpdated,
-              style: labelStyle,
-            ),
-          ),
-        ],
-      ),
     );
   }
 }
